refactor: displace schedule domain feature
