+++ src/main/java/com/example/diary/domain/schedule/controller/request/ScheduleCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.diary.controller.request;
+package com.example.diary.domain.schedule.controller.request;
 
 public record ScheduleCreateRequestDTO(String title, String content,
                                        String password, String author) {
