+++ src/main/java/com/example/diary/domain/schedule/controller/request/ScheduleUpdateRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.diary.controller.request;
+package com.example.diary.domain.schedule.controller.request;
 
 public record ScheduleUpdateRequestDTO(Long id, String title, String content,
                                        String author, String password) {
