+++ src/main/java/com/example/diary/domain/schedule/service/ScheduleServiceImpl.java
@@ -1,21 +1,17 @@
-package com.example.diary.service;
+package com.example.diary.domain.schedule.service;
 
-import com.example.diary.controller.request.ScheduleCreateRequestDTO;
-import com.example.diary.controller.request.ScheduleDeleteRequestDTO;
-import com.example.diary.controller.request.ScheduleUpdateRequestDTO;
-import com.example.diary.exception.CustomException;
-import com.example.diary.exception.ErrorCode;
-import com.example.diary.model.Schedule;
-import com.example.diary.repository.ScheduleRepository;
-import com.example.diary.service.dto.ScheduleCreateDTO;
-import com.example.diary.service.dto.ScheduleInfoDTO;
-import com.example.diary.service.dto.ScheduleUpdateDTO;
+import com.example.diary.domain.schedule.controller.request.ScheduleCreateRequestDTO;
+import com.example.diary.domain.schedule.controller.request.ScheduleDeleteRequestDTO;
+import com.example.diary.domain.schedule.controller.request.ScheduleUpdateRequestDTO;
+import com.example.diary.global.exception.CustomException;
+import com.example.diary.global.exception.ErrorCode;
+import com.example.diary.domain.schedule.model.Schedule;
+import com.example.diary.domain.schedule.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.function.Supplier;
 
 @Service
 @RequiredArgsConstructor
@@ -48,7 +44,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Override
     public ScheduleInfoDTO getScheduleById(Long id) {
         Schedule schedule = scheduleRepository.findById(id)
-                .orElseThrow(throwNotFoundException(id));
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
 
         return ScheduleInfoDTO.from(schedule);
     }
@@ -57,7 +53,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     @Transactional
     public ScheduleInfoDTO modifySchedule(ScheduleUpdateRequestDTO dto) {
         Schedule schedule = scheduleRepository.findById(dto.id())
-                .orElseThrow(throwNotFoundException(dto.id()));
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
 
         Schedule updateSchedule = schedule.update(
                 dto.title(),
@@ -66,23 +62,16 @@ public class ScheduleServiceImpl implements ScheduleService {
                 dto.password()
         );
 
-        return ScheduleInfoDTO.from(scheduleRepository.update(dto.id(),
-                ScheduleUpdateDTO.from(updateSchedule)));
+        return ScheduleInfoDTO.from(scheduleRepository.update(dto.id(), ScheduleUpdateDTO.from(updateSchedule)));
     }
 
     @Override
     @Transactional
     public void deleteById(ScheduleDeleteRequestDTO dto) {
         Schedule schedule = scheduleRepository.findById(dto.id())
-                .orElseThrow(throwNotFoundException(dto.id()));
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
 
-        schedule.delete(dto.password());
-
+        schedule.deleteSchedule(dto.password());
         scheduleRepository.deleteById(dto.id());
-    }
-
-    private static Supplier<CustomException> throwNotFoundException(Long dto) {
-        return () -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION,
-                String.format("%s는 존재하지 않습니다.", dto));
     }
 }
